chore: move hook around
